0607 what I learned in school "Spring Boot" use myBatis -> insert, update, delete with List

+++ src/main/java/kopo/poly/service/impl/StudentService.java
@@ -43,17 +43,105 @@ public class StudentService implements IStudentService {
     }
     
     @Override
+    public void insertStudentList(List<StudentDTO> pList) throws Exception {
+        log.info(this.getClass().getName() + ".insertStudentList Start!");
+        
+        pList.forEach(dto -> {
+            try {
+                this.insertStudent(dto);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+        
+        
+//        pList.forEach(dto -> {
+//            try {
+//                Optional<StudentDTO> res = Optional.ofNullable(
+//                        studentMapper.getStudent(dto)
+//                );
+//
+//                if(res.isPresent()) {
+//                    pList.remove(dto);
+//                }
+//            } catch (Exception e) {
+//                throw new RuntimeException(e);
+//            }
+//
+//        });
+//
+//        pList.forEach(dto -> {
+//            try {
+//                studentMapper.insertStudent(dto);
+//            } catch (Exception e) {
+//                throw new RuntimeException(e);
+//            }
+//        });
+        
+        
+        log.info(this.getClass().getName() + ".insertStudentList End!");
+    }
+    
+    @Override
     public void deleteStudent(StudentDTO pDTO) throws Exception {
         
         log.info(this.getClass().getName() + ".deleteStudent Start!");
         
-        studentMapper.deleteStudent(pDTO);
+        Optional<StudentDTO> res = Optional.ofNullable(
+                studentMapper.getStudent(pDTO)
+        );
         
+        if(res.isPresent()){
+            studentMapper.deleteStudent(pDTO);
+            log.info(pDTO.getUserId() + "is deleted");
+        } else {
+            log.info(pDTO.getUserId() + "is not deleted");
+        }
         
         
         
         log.info(this.getClass().getName() + ".deleteStudent is end!");
         
+    }
+    
+    @Override
+    public void deleteStudentList(List<StudentDTO> pList) throws Exception {
+        log.info(this.getClass().getName() + ".deleteStudent Start!");
+        
+        pList.forEach(dto -> {
+            try {
+                this.deleteStudent(dto);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+        
+//        pList.forEach(dto -> {
+//            try {
+//                Optional<StudentDTO> res = Optional.ofNullable(
+//                        studentMapper.getStudent(dto)
+//                );
+//
+//                if(!res.isPresent()) {
+//                    pList.remove(dto);
+//                }
+//            } catch (Exception e) {
+//                throw new RuntimeException(e);
+//            }
+//
+//
+//        });
+//
+//        pList.forEach(dto -> {
+//            try {
+//                studentMapper.deleteStudent(dto);
+//            } catch (Exception e) {
+//                throw new RuntimeException(e);
+//            }
+//        });
+        
+        
+        log.info(this.getClass().getName() + ".deleteStudent is end!");
     }
     
     @Override
@@ -67,10 +155,69 @@ public class StudentService implements IStudentService {
         ).orElseGet(ArrayList::new);
         
         
-        
-        
-
         log.info(this.getClass().getName() + ".getStudentList End!");
         return rList;
+    }
+    
+    @Override
+    public void updateStudent(StudentDTO pDTO) throws Exception {
+        log.info(this.getClass().getName() + ".updateStudent Start!");
+        Optional<StudentDTO> res = Optional.ofNullable(
+                studentMapper.getStudent(pDTO)
+        );
+        
+        if (res.isPresent()) {
+            studentMapper.updateStudent(pDTO);
+            log.info(this.getClass().getName() + ".updateStudent updating Student is success!");
+        } else {
+            log.info(this.getClass().getName() + ".updateStudent updating Student is Failed!");
+        }
+        
+        
+        log.info(this.getClass().getName() + ".updateStudent End!");
+    }
+    
+    @Override
+    public void updateStudentList(List<StudentDTO> pList) throws Exception {
+        log.info(this.getClass().getName() + ".updateStudentList Start!");
+        
+        pList.forEach(dto -> {
+            try {
+                this.updateStudent(dto);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        });
+        
+//        pList.forEach(dto -> {
+//            try {
+//                Optional<StudentDTO> res = Optional.ofNullable(
+//                        studentMapper.getStudent(dto)
+//                );
+//
+//                if(!res.isPresent()) {
+//                    pList.remove(dto);
+//                }
+//
+//
+//            } catch (Exception e) {
+//                throw new RuntimeException(e);
+//            }
+//        });
+//
+//        pList.forEach(dto -> {
+//            try {
+//                studentMapper.updateStudent(dto);
+//            } catch (Exception e) {
+//                throw new RuntimeException(e);
+//            }
+//        });
+        
+        
+        
+        
+        
+        
+        log.info(this.getClass().getName() + ".updateStudentList End!");
     }
 }
